requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
peewee-async==0.9.0
aiomysql==0.2.0
pydantic==2.5.0
psutil==5.9.6
python-dotenv==1.0.0